+++ crates/tinyfs/src/path_utils.rs
@@ -1,4 +1,7 @@
 use std::path::{Component, Path, PathBuf};
+use crate::NodePath;
+use crate::error::Error;
+use crate::error::Result;
 
 /// Strips the root component from a path, if present
 pub fn strip_root<P: AsRef<Path>>(path: P) -> PathBuf {
@@ -10,17 +13,61 @@ pub fn strip_root<P: AsRef<Path>>(path: P) -> PathBuf {
 
 /// Extracts the final component of a path as a string, if possible
 pub fn basename<P: AsRef<Path>>(path: P) -> Option<String> {
-    path.as_ref().components().last().and_then(|c| {
-        match c {
-            Component::Normal(name) => Some(name.to_string_lossy().to_string()),
-            _ => None,
-        }
+    path.as_ref().components().last().and_then(|c| match c {
+        Component::Normal(name) => Some(name.to_string_lossy().to_string()),
+        _ => None,
     })
 }
 
 /// Extracts the directory component of a path as a pathbuf, if possible
 pub fn dirname<P: AsRef<Path>>(path: P) -> Option<PathBuf> {
     path.as_ref().parent().map(|x| x.to_path_buf())
+}
+
+pub fn normalize<P: AsRef<Path>>(path: P, stack: &[NodePath]) -> Result<(usize, PathBuf)> {
+    let path = path.as_ref();
+
+    // Process components to normalize the path
+    let mut components = Vec::new();
+
+    for component in path.components() {
+        match component {
+            Component::CurDir => {} // Skip current directory components
+            Component::ParentDir => {
+                // If the last component is not a parent dir, pop it and continue
+                if let Some(Component::Normal(_)) = components.last() {
+                    components.pop();
+                    continue;
+                }
+                // Otherwise, keep the parent dir
+                components.push(component);
+            }
+            _ => components.push(component),
+        }
+    }
+
+    // Check if the path starts with a root component
+    if let Some(Component::RootDir) = components.first() {
+        return Ok((1, components.into_iter().collect()));
+    }
+
+    // Count leading parent directory components
+    let parent_count = components
+        .iter()
+        .take_while(|comp| matches!(comp, Component::ParentDir))
+        .count();
+
+    // Check if we have enough parent directories in our stack
+    if stack.len() <= parent_count {
+        return Err(Error::parent_path_invalid(path));
+    }
+
+    // Return the resulting stack size and path, skipping the parent directory components
+    // that have already been processed
+    Ok((
+        stack.len() - parent_count,
+        components.into_iter().skip(parent_count).collect(),
+    ))
 }
 
 #[cfg(test)]
@@ -53,20 +100,14 @@ mod tests {
     #[test]
     fn test_basename() {
         // Regular filename
-        assert_eq!(
-            basename("/path/to/file.txt"),
-            Some("file.txt".to_string())
-        );
-        
+        assert_eq!(basename("/path/to/file.txt"), Some("file.txt".to_string()));
+
         // Directory with trailing slash
-        assert_eq!(
-            basename("/path/to/dir/"),
-            Some("dir".to_string())
-        );
-        
+        assert_eq!(basename("/path/to/dir/"), Some("dir".to_string()));
+
         // Root directory
         assert_eq!(basename("/"), None);
-        
+
         // Empty path
         assert_eq!(basename(""), None);
     }
