+++ crates/tinyfs/src/tests/mod.rs
@@ -1,0 +1,2 @@
+mod reverse;
+mod memory;
